+++ ui/web-v2/src/grpc/account.ts
@@ -23,8 +23,7 @@ import {
   UpdateSearchFilterRequest,
   UpdateSearchFilterResponse,
   DeleteSearchFilterRequest,
-  DeleteSearchFilterResponse,
-  GetMyOrganizationsByAccessTokenRequest
+  DeleteSearchFilterResponse
 } from '../proto/account/service_pb';
 import {
   AccountServiceClient,
@@ -73,41 +72,6 @@ export function getMyOrganizations(
         request,
         getMetaData(),
         (error, response): void => {
-          if (isNotNull(error) || isNull(response)) {
-            reject(
-              new AccountServiceError(
-                extractErrorMessage(error),
-                request,
-                error
-              )
-            );
-          } else {
-            resolve({ request, response });
-          }
-        }
-      );
-    }
-  );
-}
-
-export interface GetMyOrganizationsByAccessTokenResult {
-  request: GetMyOrganizationsByAccessTokenRequest;
-  response: GetMyOrganizationsResponse;
-}
-
-export function getMyOrganizationsByAccessToken(
-  request: GetMyOrganizationsByAccessTokenRequest
-): Promise<GetMyOrganizationsByAccessTokenResult> {
-  return new Promise(
-    (
-      resolve: (result: GetMyOrganizationsByAccessTokenResult) => void,
-      reject
-    ): void => {
-      client.getMyOrganizationsByAccessToken(
-        request,
-        getMetaData(),
-        (error, response): void => {
-          console.log({ error });
           if (isNotNull(error) || isNull(response)) {
             reject(
               new AccountServiceError(
