+++ ui/web-v2/src/modules/myOrganization.ts
@@ -1,15 +1,8 @@
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
 
-import {
-  getMyOrganizations,
-  getMyOrganizationsByAccessToken
-} from '../grpc/account';
-import {
-  GetMyOrganizationsRequest,
-  GetMyOrganizationsByAccessTokenRequest
-} from '../proto/account/service_pb';
+import { getMyOrganizations } from '../grpc/account';
+import { GetMyOrganizationsRequest } from '../proto/account/service_pb';
 import { Organization } from '../proto/environment/organization_pb';
-import { AuthType } from '../proto/auth/service_pb';
 
 const MODULE_NAME = 'myOrganization';
 
@@ -18,22 +11,6 @@ export const fetchMyOrganizations = createAsyncThunk<
 >('me/fetchMyOrganizations', async () => {
   const request = new GetMyOrganizationsRequest();
   const res = await getMyOrganizations(request);
-  return res.response.toObject().organizationsList;
-});
-
-interface FetchMyOrganizationsByAccessTokenParams {
-  accessToken: string;
-}
-
-export const fetchMyOrganizationsByAccessToken = createAsyncThunk<
-  Array<Organization.AsObject>,
-  FetchMyOrganizationsByAccessTokenParams
->('me/fetchMyOrganizationsByAccessToken', async (params) => {
-  const request = new GetMyOrganizationsByAccessTokenRequest();
-  request.setAccessToken(params.accessToken);
-  request.setType(AuthType.AUTH_TYPE_GOOGLE);
-
-  const res = await getMyOrganizationsByAccessToken(request);
   return res.response.toObject().organizationsList;
 });
 
@@ -49,13 +26,5 @@ export const myOrganizationSlice = createSlice({
         myOrganization: action.payload
       };
     });
-    builder.addCase(
-      fetchMyOrganizationsByAccessToken.fulfilled,
-      (state, action) => {
-        return {
-          myOrganization: action.payload
-        };
-      }
-    );
   }
 });
